move runAndStream into a generalized helper function to enable easy local exectution in other places than just the docker builder
